(Development Version 1.1) Markdown features in help pages
The goal is to make it possible to edit the help pages from mongoDB, and at the same time make them more appealing visually.
Thus, makes room for translation's and the ability to change help pages without having to redeploy the entire app. Currently it's only available on gameFinder helppage.